+++ pkg/sql/colencoding/key_encoding.go
@@ -34,7 +34,7 @@ import (
 // See the analog in sqlbase/index_encoding.go.
 func DecodeIndexKeyToCols(
 	vecs []coldata.Vec,
-	idx uint16,
+	idx int,
 	desc *sqlbase.ImmutableTableDescriptor,
 	index *sqlbase.IndexDescriptor,
 	indexColIdx []int,
@@ -133,7 +133,7 @@ func DecodeIndexKeyToCols(
 // DecodeKeyValsToCols additionally returns whether a NULL was encountered when decoding.
 func DecodeKeyValsToCols(
 	vecs []coldata.Vec,
-	idx uint16,
+	idx int,
 	indexColIdx []int,
 	types []types.T,
 	directions []sqlbase.IndexDescriptor_Direction,
@@ -171,7 +171,7 @@ func DecodeKeyValsToCols(
 // See the analog, DecodeTableKey, in sqlbase/column_type_encoding.go.
 // decodeTableKeyToCol also returns whether or not the decoded value was NULL.
 func decodeTableKeyToCol(
-	vec coldata.Vec, idx uint16, valType *types.T, key []byte, dir sqlbase.IndexDescriptor_Direction,
+	vec coldata.Vec, idx int, valType *types.T, key []byte, dir sqlbase.IndexDescriptor_Direction,
 ) ([]byte, bool, error) {
 	if (dir != sqlbase.IndexDescriptor_ASC) && (dir != sqlbase.IndexDescriptor_DESC) {
 		return nil, false, errors.AssertionFailedf("invalid direction: %d", log.Safe(dir))
@@ -230,7 +230,7 @@ func decodeTableKeyToCol(
 		} else {
 			rkey, r, err = encoding.DecodeBytesDescending(key, nil)
 		}
-		vec.Bytes().Set(int(idx), r)
+		vec.Bytes().Set(idx, r)
 	case types.DateFamily, types.OidFamily:
 		var t int64
 		if dir == sqlbase.IndexDescriptor_ASC {
@@ -266,9 +266,7 @@ func decodeTableKeyToCol(
 // idx. An error is returned if the value's type does
 // not match the column's type.
 // See the analog, UnmarshalColumnValue, in sqlbase/column_type_encoding.go
-func UnmarshalColumnValueToCol(
-	vec coldata.Vec, idx uint16, typ *types.T, value roachpb.Value,
-) error {
+func UnmarshalColumnValueToCol(vec coldata.Vec, idx int, typ *types.T, value roachpb.Value) error {
 	if value.RawBytes == nil {
 		vec.Nulls().SetNull(idx)
 	}
@@ -301,7 +299,7 @@ func UnmarshalColumnValueToCol(
 	case types.BytesFamily, types.StringFamily, types.UuidFamily:
 		var v []byte
 		v, err = value.GetBytes()
-		vec.Bytes().Set(int(idx), v)
+		vec.Bytes().Set(idx, v)
 	case types.DateFamily, types.OidFamily:
 		var v int64
 		v, err = value.GetInt()
