+++ pkg/sql/colexec/substring_tmpl.go
@@ -99,7 +99,7 @@ func (s *substring_StartType_LengthTypeOperator) Next(ctx context.Context) colda
 	s.allocator.PerformOperation(
 		[]coldata.Vec{outputVec},
 		func() {
-			for i := uint16(0); i < n; i++ {
+			for i := 0; i < n; i++ {
 				rowIdx := i
 				if sel != nil {
 					rowIdx = sel[i]
@@ -119,7 +119,7 @@ func (s *substring_StartType_LengthTypeOperator) Next(ctx context.Context) colda
 					continue
 				}
 
-				runes := runeVec.Get(int(rowIdx))
+				runes := runeVec.Get(rowIdx)
 				// Substring start is 1 indexed.
 				start := int(startVec[rowIdx]) - 1
 				length := int(lengthVec[rowIdx])
@@ -142,7 +142,7 @@ func (s *substring_StartType_LengthTypeOperator) Next(ctx context.Context) colda
 				} else if start > len(runes) {
 					start = len(runes)
 				}
-				outputCol.Set(int(rowIdx), runes[start:end])
+				outputCol.Set(rowIdx, runes[start:end])
 			}
 		},
 	)
